+++ CHANGELOG.md
@@ -33,6 +33,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Bug Fixes
 
 - Accounts not refreshing in the forms.
+- Edit Transaction Tabs not disabled.
+- Edit Form Title.
 
 ---